# Reduce memory allocations in decode_cef processor